% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktm_rain17.R
\docType{data}
\name{ktm_rain17}
\alias{ktm_rain17}
\title{Daily rainfall data of Kathmandu for 2017.}
\format{
A data frame with 365 rows and 2 variables:
\describe{
  \item{\code{TIME}}{Date in YYYY-MM-DD format.}
  \item{\code{rainfall}}{double rainfall in millimeters}
}
}
\source{
\url{http://www.mfd.gov.np/city?id=31/}
}
\usage{
ktm_rain17
}
\description{
This file contains daily rainfall data of Kathmandu. The source of this data is 'Government of Nepal, Department of Hydrology and Meteorology'.
}
\keyword{datasets}
