% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nepa.R
\docType{data}
\name{nepa}
\alias{nepa}
\title{Dendrometer data from Kathmandu}
\format{
A data frame with 14534 rows and 3 variables:
\describe{
  \item{\code{Time}}{datetime time of data recording}
  \item{\code{T2}}{double reading for first tree}
  \item{\code{T3}}{double reading for second tree}
}
}
\usage{
nepa
}
\description{
Dendrometer data from three Chir pine trees collected in hourly resolution for 2 years.
}
\keyword{datasets}
