% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nepa17.R
\docType{data}
\name{nepa17}
\alias{nepa17}
\title{Dendrometer data of Kathmandu for 2017}
\format{
A data frame with 8753 rows and 3 variables:
\describe{
  \item{\code{Time}}{datetime time of data recording}
  \item{\code{T2}}{double reading for first tree}
  \item{\code{T3}}{double reading for second tree}
}
}
\usage{
nepa17
}
\description{
Dendrometer data from three Chir pine tree collected in hourly resolution for 2017.
}
\keyword{datasets}
