% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.dendrometer.R
\name{read.dendrometer}
\alias{read.dendrometer}
\title{Reading dendrometer data.}
\usage{
read.dendrometer(file, sep = NULL, dec = NULL)
}
\arguments{
\item{file}{string file name or path of the file.}

\item{sep}{string the separator of the files. Only if they are different than the standard separators such as  tab for .txt file and comma for .csv file.}

\item{dec}{the character used in the file for decimal points.}
}
\value{
A dataframe with the dendrometer data:
}
\description{
This function reads dendrometer data from .csv or .txt or .xlsx files. This function automatically recognizes the date time format of the first column of the data frame and changes it to "yyyy-mm-dd HH:MM:SS format".
}
